+++ Typin/Directives/DefaultDirective.cs
@@ -6,10 +6,10 @@
     using Typin.Console;
 
     /// <summary>
-    /// Normally if application rans in interactive mode, an empty line does nothing; but [default] will override this behaviour, executing a root (empty) command or scoped command without arguments.
+    /// Normally when application runs in interactive mode, an empty line does nothing; but [default] will override this behaviour, executing a root or scoped command.
     /// </summary>
     [ExcludeFromCodeCoverage]
-    [Directive("default", Description = "Executes a root (empty) command or scoped command without arguments (parameters and options).", InteractiveModeOnly = true)]
+    [Directive(BuiltInDirectives.Default, Description = "Executes a root or scoped command.", InteractiveModeOnly = true)]
     public sealed class DefaultDirective : IDirective
     {
         /// <inheritdoc/>
@@ -26,14 +26,6 @@
         /// <inheritdoc/>
         public ValueTask HandleAsync(IConsole console)
         {
-            //TODO: maybe make [default] -h etc forbidden
-            //bool isInteractiveMode = _cliContext.IsInteractiveMode;
-            //string scope = _cliContext.Scope;
-            //CommandInput input = _cliContext.CurrentInput;
-
-            // if (input.IsDefaultCommandOrEmpty)
-            //ContinueExecution = true;
-
             return default;
         }
     }