+++ Typin/Input/CommandInput.cs
@@ -51,7 +51,7 @@
         public bool IsVersionOptionSpecified => Options.Any(o => o.IsVersionOption);
 
         /// <summary>
-        /// Whether command input is default command or empty (no command name, no options, no parameters, and no directives other than [interactive]_.
+        /// Whether command input is default command or empty (no command name, no options, no parameters, and no directives other than [interactive].
         /// </summary>
         public bool IsDefaultCommandOrEmpty { get; }